CHANGE TIP REPOSITORY

+++ src/main/java/ec/gob/conagopare/sona/modules/content/models/Tip.java
@@ -56,7 +56,7 @@ public class Tip extends Auditable implements Persistable<UUID>, PurgableStored 
     private Double averageRate;
 
     @Transient
-    private Integer totalRate;
+    private Long totalRate;
 
     @Override
     public String[] filesFullPaths() {
